Changed setup script to tolerate final QC warnings
This takes advantage of the new --exit-on-warn parameter to qcquiet.sh
and the change to qc_bam.sh to make PairedReadCount and Read1And2Counts
into warnings.  Before this change, if the input BAM had unusual paired
read counts, then the whole pipeline would fail in the final QC.  Now,
those will be reported as warnings.  It can still fail on bigger
problems.